Models/requirements.txt
langchain
langchain-core
langchain_community

langchain-openai
openai

langchain-anthropic

langchain-google-genai
google-generativeai

langchain-huggingface
transformers
huggingface-hub
sentence-transformers

python-dotenv
scikit-learn
numpy
jinja2
torch 